+++ backend/packages/caching/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @cat-factory/caching
 
+## 0.6.16
+
+### Patch Changes
+
+- Updated dependencies [7ee2530]
+  - @cat-factory/kernel@0.112.1
+
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies [f25d5e2]
+  - @cat-factory/kernel@0.112.0
+
+## 0.6.14
+
+### Patch Changes
+
+- @cat-factory/kernel@0.111.1
+
+## 0.6.13
+
+### Patch Changes
+
+- Updated dependencies [63f7881]
+  - @cat-factory/kernel@0.111.0
+
 ## 0.6.12
 
 ### Patch Changes