+++ frontend/app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/app
 
+## 0.109.2
+
+### Patch Changes
+
+- Updated dependencies [9aa9e19]
+  - @cat-factory/contracts@0.121.1
+
+## 0.109.1
+
+### Patch Changes
+
+- Updated dependencies [63f7881]
+  - @cat-factory/contracts@0.121.0
+
 ## 0.109.0
 
 ### Minor Changes